backend/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.7.14 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
fastapi==0.115.14 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.25.1 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.40.3 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==3.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.70.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.12" and python_version < "4.0"
protobuf==6.31.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.30.6 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.12" and python_version < "4.0"
